feat(frontend): integrate TanStack Query v5 with openapi-react-query
- Add @tanstack/react-query, @tanstack/react-query-devtools, openapi-react-query
- Upgrade openapi-fetch to 0.17 (peer dep requirement)
- Add QueryClient setup with isServer branch (query-client.ts)
- Add  client for type-safe useQuery (lib/api/react-query.ts)
- Create unified Providers component (QueryClientProvider + ChakraProvider)
- Migrate artifact-viewer-section.tsx from useEffect+fetch to useQuery
- Extract RepositoriesList client component with .useQuery
- Convert repositories page to prefetch + HydrationBoundary pattern

Closes #64
