+++ boardflow/package.json
@@ -13,9 +13,12 @@
   "dependencies": {
     "@chakra-ui/react": "^3.35.0",
     "@emotion/react": "^11",
+    "@tanstack/react-query": "^5.100.9",
+    "@tanstack/react-query-devtools": "^5.100.9",
     "lucide-react": "^0.500",
     "next": "^15",
-    "openapi-fetch": "^0.13",
+    "openapi-fetch": "^0.17.0",
+    "openapi-react-query": "^0.5.4",
     "react": "^19",
     "react-dom": "^19"
   },
