+++ boardflow/src/components/providers.tsx
@@ -1,0 +1,16 @@
+"use client"
+
+import { QueryClientProvider } from "@tanstack/react-query"
+import { ReactQueryDevtools } from "@tanstack/react-query-devtools"
+import { ChakraProvider, defaultSystem } from "@chakra-ui/react"
+import { getQueryClient } from "@/lib/query-client"
+
+export function Providers({ children }: { children: React.ReactNode }) {
+  const queryClient = getQueryClient()
+  return (
+    <QueryClientProvider client={queryClient}>
+      <ChakraProvider value={defaultSystem}>{children}</ChakraProvider>
+      <ReactQueryDevtools />
+    </QueryClientProvider>
+  )
+}
