+++ boardflow/src/lib/api/react-query.ts
@@ -1,0 +1,4 @@
+import createClient from "openapi-react-query"
+import { apiClient } from "./client"
+
+export const $api = createClient(apiClient)
